Atualização das classes referentes ao experimento.

+++ src/java/br/ufjf/pgcc/plscience/dao/TavernaWorkflowDAO.java
@@ -49,7 +49,7 @@ public class TavernaWorkflowDAO extends GenericDAO {
         finish();        
     }
     
-    public List<TavernaWorkflow> getExperimentWorkflows(Long experimentId) {
+    public List<TavernaWorkflow> getExperimentWorkflows(Integer experimentId) {
         Query query = getEntityManager().createQuery("SELECT t FROM TavernaWorkflow t WHERE t.experiment.id = :p_experiment_id ORDER BY t.createdAt DESC");
         query.setParameter("p_experiment_id", experimentId);
         List<TavernaWorkflow> workflows = query.getResultList();
